backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
httpx==0.25.2
pyyaml==6.0.1
python-multipart==0.0.6
apscheduler==3.10.4
python-telegram-bot==20.7
aiofiles==23.2.1
beautifulsoup4==4.12.3
lxml==5.1.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-dateutil==2.8.2
anthropic==0.40.0
